Move common logic into ProcessInfo & ProcessItem
Changes:
- ProcessInfo responsible for the depositOutputs, retieveInpouts & returnInputs as logic moved out of WorkshopProcess
- ProcessItem handles depositing items into a Settlement
- ManufactureProcess & WorkshopProcess uses the delegated logic
- FoodProduction uses the new shared logic from ProcessInfo

Part of #871
