+++ mars-sim-core/src/main/java/com/mars_sim/core/process/ProcessItem.java
@@ -8,7 +8,20 @@ package com.mars_sim.core.process;
 
 import java.io.Serializable;
 
+import com.mars_sim.core.Simulation;
+import com.mars_sim.core.equipment.BinFactory;
+import com.mars_sim.core.equipment.EquipmentFactory;
+import com.mars_sim.core.equipment.EquipmentType;
+import com.mars_sim.core.goods.Good;
+import com.mars_sim.core.goods.GoodsUtil;
+import com.mars_sim.core.logging.SimLogger;
+import com.mars_sim.core.resource.ItemResourceUtil;
 import com.mars_sim.core.resource.ItemType;
+import com.mars_sim.core.resource.Part;
+import com.mars_sim.core.structure.Settlement;
+import com.mars_sim.core.vehicle.Vehicle;
+import com.mars_sim.core.vehicle.VehicleFactory;
+import com.mars_sim.core.vehicle.VehicleType;
 
 /**
  * A Process input or output item.
@@ -18,6 +31,8 @@ public class ProcessItem implements Serializable {
 
 	/** default serial id. */
 	private static final long serialVersionUID = 1L;
+
+    private static final SimLogger logger = SimLogger.getLogger(ProcessItem.class.getName());
 
 	// Data members
 	private String name;
@@ -47,6 +62,92 @@ public class ProcessItem implements Serializable {
 
 	public double getAmount() {
 		return amount;
+	}
+
+    /**
+     * Deposits the item into the settlement.
+     * @param settlement
+     * @param context The process driving the deposit
+     * @param updateGoods Updated the Goods manager for this item
+     */
+    void deposit(Settlement settlement, ProcessInfo context, boolean updateGoods) {
+		int outputId = -1;
+		double outputAmount = amount;
+		switch(type) {
+			case AMOUNT_RESOURCE: {
+
+				// Produce amount resources.
+				outputId = id;
+				double capacity = settlement.getAmountResourceRemainingCapacity(outputId);
+				if (outputAmount> capacity) {
+					double overAmount = amount - capacity;
+					logger.severe(settlement, "Process " + context.getName() + " Not enough storage capacity to store " 
+						+ Math.round(overAmount * 10.0)/10.0 + " kg " + name
+							+ " from '" + name + "'.");
+					outputAmount = capacity;
+				}
+				settlement.storeAmountResource(outputId, outputAmount);
+			} break;
+
+			case PART: {
+				// Produce parts.
+				outputId = id;
+				Part part = ItemResourceUtil.findItemResource(outputId);
+				int num = (int)outputAmount;
+				double mass = num * part.getMassPerItem();
+				double capacity = settlement.getCargoCapacity();
+				if (mass <= capacity) {
+					settlement.storeItemResource(outputId, num);
+				}
+				else {
+					outputId = -1;
+				}
+			} break;
+
+			case EQUIPMENT: {
+				// Produce equipment.
+				var equipmentType = EquipmentType.convertName2Enum(name);
+				outputId = EquipmentType.getResourceID(equipmentType);
+				int number = (int) outputAmount;
+				for (int x = 0; x < number; x++) {
+					EquipmentFactory.createEquipment(equipmentType, settlement);
+				}
+			} break;
+
+			case BIN: {
+				// Produce bins.
+				int number = (int) outputAmount;
+				for (int x = 0; x < number; x++) {
+					BinFactory.createBins(name, settlement);
+				}
+			} break;
+		
+			case VEHICLE: {
+				// Produce vehicles.
+				int number = (int) outputAmount;
+				var unitMgr = Simulation.instance().getUnitManager();// Don't like this
+				for (int x = 0; x < number; x++) {
+					Vehicle v = VehicleFactory.createVehicle(unitMgr, settlement, name);
+
+					outputId = VehicleType.getVehicleID(v.getVehicleType());
+				}
+			} break;
+		}
+
+		// Record goods benefit
+		if (updateGoods) {
+			if (outputId >= 0) {
+				settlement.addOutput(outputId, outputAmount, context.getWorkTimeRequired());
+			}
+
+			Good good = GoodsUtil.getGood(name);
+			if (good == null) {
+				logger.severe(name + " is not a good.");
+			}
+			else
+				// Recalculate settlement good value for the output item.
+				settlement.getGoodsManager().determineGoodValue(good);
+		}
 	}
 
 	@Override
